feat(project): adds production compose

+++ scripts/setup.dev.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 # Get useful paths
 SCRIPT_PATH="$(realpath "${BASH_SOURCE[0]}")"
@@ -19,8 +19,15 @@ if [ $? -ne 0 ]; then
 fi
 
 # Set up environment variables
-ENV_EXAMPLE_FILE="$COLLABST_DIR/.env.example"
-ENV_FILE="$COLLABST_DIR/.env"
+ENV_EXAMPLE_FILE="$COLLABST_DIR/config/env/.env.dev.example"
+ENV_FILE="$COLLABST_DIR/config/env/.env"
+
+if [ ! -f "$ENV_EXAMPLE_FILE" ]; then
+    echo "Missing env template at $ENV_EXAMPLE_FILE"
+    exit 1
+fi
+
+mkdir -p "$(dirname "$ENV_FILE")"
 
 cp "$ENV_EXAMPLE_FILE" "$ENV_FILE"
